requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
pandas==2.1.3
openpyxl==3.1.2
PyYAML==6.0.1